+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -9,7 +9,7 @@
 
 int main(void)
 {
-	int a = 1, b = 2, c, i, sum = 3;
+	long int a = 1, b = 2, c, i, sum = 3;
 
 	while (c <= 4000000)
 	{
@@ -20,7 +20,7 @@ int main(void)
 		a = b;
 		b = c;
 	}
-	printf("%d\n", sum);
+	printf("%ld\n", sum);
 
 	return (0);
 }
